Update core module dependencies to latest alpha versions (#731)
# Description
Update core_modules to validate latest alpha version compat.

# Issues
Blacklists an onnxruntime version to mitigate
https://github.com/microsoft/onnxruntime/issues/23957

# Other Notes
Based on [this run from an old tag
failing](https://github.com/NeonGeckoCom/NeonCore/actions/runs/13816049514/job/38649389777),
it appears that the failures are related to some change in GitHub
Actions runners or some low-level dependency change.

Cause for Pi image failures is unknown, but updating to a newer CPU arch
in the shared action appears to resolve the issue. Note that the
previously spec'd A53 was used in the RPi3 and the now spec'd A76 is
used in the RPi5

Issue was traced back to a new onnxruntime release